requirements.txt
streamlit>=1.30.0
pandas>=2.1.0
numpy>=1.26.0
plotly>=5.18.0
yfinance>=0.2.33
praw>=7.7.1
vaderSentiment>=3.3.2
python-dotenv>=1.0.0
watchdog>=3.0.0
feedparser